# openapi.yaml
openapi: 3.1.0
info:
  title: PulseBoard Ingest API
  version: 1.0.0
  description: >-
    Public HTTP API for sending product events to PulseBoard and stitching
    anonymous visitors to known users. Events are validated, deduplicated by
    idempotency key, and written to ClickHouse.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/identify:
    post:
      summary: Stitch an anonymous id to a user id
      description: >-
        Records `anonymousId -> userId` so historical anonymous events resolve
        to the same actor in analytics. Idempotent per anonymousId.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [anonymousId, userId]
              properties:
                anonymousId: { type: string }
                userId: { type: string }
      responses:
        '200': { description: Alias recorded }
        '400': { description: Validation failure or equal ids }
        '401': { description: Missing or invalid bearer token }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: INGEST_API_KEY as a bearer token.
...